Basic SecureString support
Even Microsoft's own password-related functions do not support
SecureString in ANY version of the .NET Framework, so we end up
converting to string in memory at some point.

This way we can at least at some point in the future use unsafe code to
access its bytes in the memory directly for clients using that
interface.