Full (and tested) HMAC authentication support
We either create a double-wide key or derive a double-wide key from the
user's password and split that in half, one key used for encryption and
the other for authentication. (Even though it would probably be OK to
use the same key for both given the differences in nature between AES
and HMAC, but this is better).

We use Encrypt-then-HMAC to authenticate the (iv, ciphertext) tuple with
HMAC with the second key, and store the resulting hash in the blob
alongside the iv and the ciphertext. This HMAC is validated _before_
decryption when an attempt to retrieve encrypted data is made the next
go around.