Correctly dispose resources in case of TamperedCipherTextException
`SecretsManager.Encrypt()` would previously construct an instance of a
`SymmetricAlgorithm` prior to validating the HMAC; if a mismatch were detected,
a `TamperedCipherTextException` would be thrown and the `SymmetricAlgorithim`
instance would not be disposed (it couldn't be instantiated in a using block due
to the differences in supported APIs across target .NET Framework versions).

The code now validates the HMAC before constructing the AES `SymmetricAlgorithm`
thereby addressing the issue.