Protect against Newtonsoft.Json vulnerability GHSA-5crp-9r3c-p9vr
Per [0], dependency `Newtonsoft.Json`, used to serialize or deserialize the
vault to/from disk, has a known issue that can allow abuse of CPU and/or RAM by
an attacker that can control what gets passed in for deserialization.

While the vault is technically secure and not subject to user-controlled input,
out of an abundance of caution a cap is set on the `JsonSerializerSettings`
`MaxDepth` property and it is passed to all serialization/deserialization call
sites.

[0]: https://github.com/advisories/GHSA-5crp-9r3c-p9vr