Simplify conditional code by dropping support for legacy targets
The project is now only built for .NET Standard 2.0 (to support .NET 4.x and UWP
10.0.16299+) and .NET 6 (for everyone else). This lets us simplify the
conditional code considerably.

.NET Core 2.x web apps (already EOL) are still supported via .NET Standard 2.0,
as are .NET Core 3.x web apps (EOL: Dec 2022).

The tests are now run against .NET 4.8 and .NET 6. The client is only built for
.NET 6 as it is not used in production, only on development machines (where it
isn't unfair to put in place such a runtime requirement).