Harden the default VaultVersionPolicy
After the leeway provided during the rollout of vault versioning and vault
version 2, now tighten the default `VaultVersionPolicy` to prevent schema
downgrade attacks.

The `SecureStore` library now defaults the static property
`SecretsManager.VaultVersionPolicy` to `VaultVersionPolicy.Strict` which forbids
upgrades across major vault versions to protect against schema downgrade attacks
which can be used to bypass stricter validation/security introduced in more
recent vault versions.

The companion CLI app overrides this to allow upgrades by default to afford a
middle ground where upgrades are allowed under the scrutiny and supervision of a
developer/deployer with the commit bit, but protect against attacker-originated
vault downgrade attacks in-prod once deployed and live.

A `VaultVersionException` is thrown if an attempt to load a vault with a lower
major version is attempted while `VaultVersionPolicy.Strict` is in place.